requirements.txt
asgiref==3.11.0
boto3==1.42.16
botocore==1.42.16
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
idna==3.11
jmespath==1.0.1
mysqlclient==2.2.7
pillow==12.0.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
wheel==0.45.1
Pillow==12.0.0